Fixed global variables, fixed cooked food plate order
Fixed global variables being confusing in restaurant script. Problem was the script was attached to the scene so there was a global script and a script that was attached to the scene. Solved it by making a new global script to store global variables and methods.

Also reworked the food ready on the bar. Before a plate was chosen before the food was ready, leading to gaps as the food was cooking. Fixed this by having the food cook first(timer), and then choosing an available plate. Can't believe I didn't think of that when I made it.